requirements.txt
streamlit
sentence-transformers
Pillow
torch
